+++ examples/tracing/jaeger/go/rideshare/loadgen.go
@@ -22,6 +22,7 @@ var hosts = []string{
 	"us-east-java",
 	"eu-north-java",
 	"ap-south-java",
+	"eu-north-java-instrumentation",
 }
 
 var vehicles = []string{