+++ pkg/analytics/analytics_test.go
@@ -41,7 +41,7 @@ func (*mockStatsProvider) AppsCount() int { return 0 }
 
 var _ = Describe("analytics", func() {
 	gracePeriod = 100 * time.Millisecond
-	uploadFrequency = 200 * time.Millisecond
+	oldMetricsUploadFrequency = 200 * time.Millisecond
 	snapshotFrequency = 200 * time.Millisecond
 
 	testing.WithConfig(func(cfg **config.Config) {
@@ -69,7 +69,7 @@ var _ = Describe("analytics", func() {
 
 					httpServer := httptest.NewServer(myHandler)
 					defer httpServer.Close()
-					url = httpServer.URL + "/api/events"
+					host = httpServer.URL
 
 					s, err := storage.New(storage.NewConfig(&(*cfg).Server), logrus.StandardLogger(), prometheus.NewRegistry(), new(health.Controller))
 					Expect(err).ToNot(HaveOccurred())
@@ -107,7 +107,7 @@ var _ = Describe("analytics", func() {
 
 					httpServer := httptest.NewServer(myHandler)
 					defer httpServer.Close()
-					url = httpServer.URL + "/api/events"
+					host = httpServer.URL
 
 					s, err := storage.New(storage.NewConfig(&(*cfg).Server), logrus.StandardLogger(), prometheus.NewRegistry(), new(health.Controller))
 					Expect(err).ToNot(HaveOccurred())