+++ packages/pyroscope-flamegraph/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.21.4](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.21.3...@pyroscope/flamegraph@0.21.4) (2022-09-06)
+
+
+### Bug Fixes
+
+* **flamegraph:** add color to tooltip ([#1468](https://github.com/pyroscope-io/pyroscope/issues/1468)) ([1c29ef6](https://github.com/pyroscope-io/pyroscope/commit/1c29ef6328fb4f01389560a4d53f729bbca88ff3))
+
+
+
+
+
+## [0.21.3](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.21.2...@pyroscope/flamegraph@0.21.3) (2022-09-06)
+
+
+### Bug Fixes
+
+* **flamegraph:** table width ([#1466](https://github.com/pyroscope-io/pyroscope/issues/1466)) ([a60f608](https://github.com/pyroscope-io/pyroscope/commit/a60f608475a9caea0e1910f30d1c663fd38cee96))
+
+
+
+
+
+## [0.21.2](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.21.1...@pyroscope/flamegraph@0.21.2) (2022-09-06)
+
+
+### Bug Fixes
+
+* **flamegraph:** table width ([#1463](https://github.com/pyroscope-io/pyroscope/issues/1463)) ([f19b8ac](https://github.com/pyroscope-io/pyroscope/commit/f19b8ac778452ea86261563aa1d165dbc2d089e7))
+
+
+
+
+
 ## [0.21.1](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/flamegraph@0.21.0...@pyroscope/flamegraph@0.21.1) (2022-09-05)